Refactored folder name and imports, jsdoc updates
