+++ frontend-react/src/components/Content/DirectoryGenerationTools.tsx
