+++ frontend-react/src/components/Content/GeneratedSideNav.test.tsx
