+++ frontend-react/src/components/Content/GeneratedSideNav.tsx
