+++ frontend-react/src/components/Content/MarkdownDirectory.tsx
@@ -4,15 +4,13 @@ import { Route } from "react-router-dom";
 
 import DirectoryAsPage from "./DirectoryAsPage";
 
-/* Used to instantiate a set of static pages, like BuiltForYouIndex
- * or HowItWorks */
-
 type ContentElement = () => JSX.Element;
 
-export interface MarkdownPageProps {
-    directories: MarkdownDirectory[];
-}
-
+/** A base type that holds directory information
+ *
+ * @property title
+ * @property slug
+ * @property desc */
 export abstract class ContentDirectory {
     title: string = "";
     slug: string = "";
@@ -23,12 +21,13 @@ export abstract class ContentDirectory {
         this.desc = desc;
     }
 }
-
-/** Creates a directory (page) consisting of one ore many elements to render in order
+/** Creates a backwards-compatible method of rendering old React elements
+ * as pages until converted to markdown
  *
- * @param title {string} Value displayed in GeneratedSideNav
- * @param slug {string} the url slug to navigate to
- * @param element {ContentElement[]} one or more elemnets to render on a given page
+ * @property title
+ * @property slug
+ * @property desc
+ * @property element - Element to render
  */
 export class ElementDirectory extends ContentDirectory {
     element: ContentElement;
@@ -42,10 +41,14 @@ export class ElementDirectory extends ContentDirectory {
         this.element = element;
     }
 }
-
-/* Used to create objects that hold pointers to markdown directories and the
+/** Used to create objects that hold pointers to markdown directories and the
  * info needed to query them. This is because we cannot access the filesystem
- * at runtime */
+ * at runtime
+ *
+ * @property title
+ * @property slug
+ * @property desc
+ * @property files - markdown files to render */
 export class MarkdownDirectory extends ContentDirectory {
     files: module[];
     constructor(title: string, slug: string, files: module[], desc?: string) {
@@ -53,7 +56,7 @@ export class MarkdownDirectory extends ContentDirectory {
         this.files = files;
     }
 }
-
+/** Takes a `ContentDirectory` and returns a react-router `Route` */
 export const GeneratedRoute = ({ dir }: { dir: ContentDirectory }) => {
     if (dir instanceof MarkdownDirectory) {
         return (
@@ -79,7 +82,8 @@ export const GeneratedRoute = ({ dir }: { dir: ContentDirectory }) => {
         );
     }
 };
-
+/** Takes a `ContentDirectory[]` and generates a React Fragment containing
+ * each directory's `GeneratedRoute` */
 export const GeneratedRouter = ({
     directories,
 }: {
