+++ frontend-react/src/components/Content/StaticPagesFromDirectories.tsx
