+++ frontend-react/src/pages/admin/InternalUserGuides.tsx
@@ -1,6 +1,6 @@
-import { MarkdownDirectory } from "../../components/Markdown/MarkdownDirectory";
+import { MarkdownDirectory } from "../../components/Content/MarkdownDirectory";
 import markdownPagesGuide from "../../content/internal-user-guides/make-markdown-pages.md";
-import StaticPagesFromDirectories from "../../components/Markdown/StaticPagesFromDirectories";
+import StaticPagesFromDirectories from "../../components/Content/StaticPagesFromDirectories";
 
 export const InternalUserGuidesDirectory = [
     new MarkdownDirectory("Create markdown pages", "create-markdown-pages", [
